+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [3.1.0](https://github.com/sketch7/signalr-client/compare/3.0.0...3.1.0) (2022-11-22)
+
+### Features
+
+- **hub connection:** ability to configure microsoft `HubConnection` directly via `configureSignalRHubConnection`
+
+### Bug Fixes
+
+- **hub connection:** correctly close the websocket connection when `disconnect` is called right after `connect`
+- **hub connection:** `connectionState$` `connecting` will also be set when connecting not only when reconnecting
+
 ## [3.0.0](https://github.com/sketch7/signalr-client/compare/2.0.0...3.0.0) (2021-04-30)
 
 ### Features